notebooks/requirements.txt
joblib
textblob
pandas-profiling
vaderSentiment
spacy
jellyfish
langdetect
